Change data type

+++ src/main/java/pl/agilevision/hardware/um7/UM7Client.java
@@ -20,7 +20,7 @@ public interface UM7Client {
 
   boolean isConnected();
 
-  int readByte();
+  byte readByte();
 
   UM7Packet readPacket() throws DeviceConnectionException;
